+++ routers/user.js
@@ -22,7 +22,6 @@ function* list(next) {
     const {number, page} = this.params;
     const userList = yield getUserList(number, page);
     if ("json" === this.way) {
-      this.set("Content-Type", "application/json; charset=utf-8");
       this.body = this.success(userList);
     }
   }catch(err) {
@@ -40,7 +39,6 @@ function* permit(next) {
     const {mail} = this.params;
     yield obmitLoginPermit(mail);
     if ("json" === this.way) {
-      this.set("Content-Type", "application/json; charset=utf-8");
       this.body = this.success();
     }
   } catch (err) {
@@ -62,13 +60,27 @@ function* login(next) {
     }
     const sessionId = yield loginSystem(mail, permit);
     if ("json" === this.way) {
-      this.set("Content-Type", "application/json; charset=utf-8");
       this.body = this.success(sessionId);
     }
     this.forceSign(sessionId);
   } catch (err) {
     this.error = err;
-    console.log(err);
+  }
+
+  return yield next;
+}
+
+function* skipLogin(next) {
+  if (this.error) {
+    return yield next;
+  }
+
+  try {
+    if ("web" === this.way) {
+
+    }
+  } catch (err) {
+    this.error = err;
   }
 
   return yield next;
@@ -78,4 +90,5 @@ module.exports = {
   list,
   permit,
   login,
+  skipLogin,
 };